Debugging port. It seems the last commit fixed the API bugs. However, the fitting speed isn't as improved as I would have hoped. Turns out the improvements in PyMC4 are not the silver bullet I would have expected. :'-(  So, I'm trying to make some adjustments to the elong_logp function in hopes I can identify the bottleneck. First I'm checking to see if the data type for the normalization vector has an impact (changing from int64 --> int32).

+++ liet/rnap_lib_LIET_class.py
@@ -434,7 +434,7 @@ class LIET:
 #                    + self._p['tI']**2), 
 #                    self._p['mT'] + _n*self._p['sT']
 #                ]))
-                _x = tt.arange(_min, _max, dtype="int64")
+                _x = tt.arange(_min, _max, dtype="int32")
 
                 _norm_array = (
                     _emg_cdf(_x, mu=mL, sigma=sL, tau=tI)
